requirements.txt
Flask
PyMuPDF
python-docx
openpyxl
pandas
xlrd